Close #3: Improve download buttons
Particularly with regard to touch devices, the download buttons are
rather meaningless, because users have no way to tell which file
they're going to download and how large it is. Furthermore, PNG images
aren't good enough anymore with regard to high-resolution devices.

Therefore we dump the download button images altogether, and instead
have a simple button with an appropriate label. We leave it to the
webmasters to style the download buttons as they see fit.